+++ regen/compute/prefix-scan.cpp
